+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.5.0 (Preview)
+
+* support aggregation metric alias
+
 ### 0.4.0
 
 * support filter function against ES version >= 5.0